refactored kmcmultiple and moved out some classes
