+++ netbeans/libxtp/nbproject/private/private.xml
@@ -10,14 +10,14 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/home/jens/votca_git/src/xtp/src/libxtp/calculators/neighborlist.h</file>
+            <file>file:/home/jens/votca_git/src/xtp/include/votca/xtp/bulkesp.h</file>
             <file>file:/home/jens/votca_git/src/xtp/src/libxtp/gnode.cc</file>
+            <file>file:/home/jens/votca_git/src/xtp/src/libxtp/calculators/kmcmultiple.cc</file>
             <file>file:/home/jens/votca_git/src/xtp/src/libxtp/calculators/kmcmultiple.h</file>
-            <file>file:/home/jens/votca_git/src/xtp/src/libxtp/diis.cc</file>
+            <file>file:/home/jens/votca_git/src/xtp/include/votca/xtp/chargecarrier.h</file>
             <file>file:/home/jens/votca_git/src/xtp/src/libxtp/calculators/kmclifetime.cc</file>
             <file>file:/home/jens/votca_git/src/xtp/include/votca/xtp/kmccalculator.h</file>
             <file>file:/home/jens/votca_git/src/xtp/include/votca/xtp/gnode.h</file>
-            <file>file:/home/jens/votca_git/src/xtp/src/libxtp/calculatorfactory.cc</file>
             <file>file:/home/jens/votca_git/src/xtp/src/libxtp/kmccalculator.cc</file>
             <file>file:/home/jens/votca_git/src/xtp/src/libxtp/calculators/kmclifetime.h</file>
         </group>
